refractoring with clean some layout

+++ db/migrate/20110228170637_change_price_type.rb
@@ -1,7 +1,7 @@
 class ChangePriceType < ActiveRecord::Migration
   def self.up
     remove_column :posts, :price
-    add_column :posts, :price, :integer
+    add_column :posts, :price, :float
   end
 
   def self.down
